Destroy io_context before waiting for other threads
This change makes no difference for Emilua core.

While developing a plugin to communicate with Qt's QCoreApplication
thread it became apparent that a service's destructor attached to
io_context can be an useful technique to interrupt extra threads. To
make the life of some plugin developers easier I'm pushing this commit.

This behavior already existed for io_context objects of other threads
(they were destroyed before the thread joining phase), and there's no
reason for making the behavior of the first io_context special in this
regards. Now the behavior is unified which is good anyways.