Add missing lua_checkstack() on variadic arguments for fiber_resume()
lua_checkstack() on variadic argumens is user's responsibility, but only
for the arguments he pushes himself. We still push a few values after
the user is done with his own arguments. That's the check that was
missing.

Anyhow, we still don't have custom memory allocators, so allocation can
never fail and the fixed bug wouldn't even manifest anyway. Fixing it
anyway as support for fallible memory allocators are part of the
roadmap.