Refactor build into library + executable
Only main.ypp is not included as part of the library sources. This
change should pave the way for custom builds that embed the new
libemilua into a custom executable that also contains a target Lua
program built-in.

Therefore the motivating factor for this change is to make it easier to
employ new approaches for packaging and deploying programs.

A package manager that can automatically make use of libemilua should
also be written later down the road.