Add setxid logic to synchronize credentials among auxiliary processes
The auxiliary single-threaded process forked at startup to spawn further
processes without the need to call exec() at the end could be used to
bypass any drop of credentials performed through calls such as
system.setresuid().

This patch fixes the problem. Now such calls will only return to the Lua
program once we receive confirmation that the same operation was
successfully performed in auxiliary processes (which happens to be only
1 extra process today).