Be explicit about mutex lock state during plugins initialization
As I was creating a plugin for gobject-introspection, I've learned about
the value of passing the locks down the chain. shared_mutex is not a
recursive mutex. One cannot freely assume it can acquire a lock if
current thread already owns it.

For the gobject-introspection plugin, I needed a lock to pre-populate
error_category objects in rdf_ec_cache_registry from init_appctx().