Fix argv overrun
Although uncommon, &(char*){NULL} is a valid value for argv (i.e. when
argc is zero).

The original loop was actually correct. Commit
529c32a2271d89096897daf87d9cf9ac79053aca broke the code when it changed
the init-clause, but left the cond-expression untouched.

The bug isn't really important today. However there are plans to make
Emilua able to create suid binaries. So this fix is important for the
future where it'll matter.