Fix FreeBSD environ handling
FreeBSD is very picky about how we manipulate environ. First of all,
clearenv() and other functions to interact with environ will pollute
stdout/stderr if they detect any environment variable with no equals
sign (a "corrupt" environment).

Second of all, it'll complain if we set environ[0] to NULL:
https://lists.freebsd.org/pipermail/freebsd-stable/2008-June/043113.html

So this commit unsets environ[0]. To deal with the many problems of
clearenv() and other functions that may pollute stdout/stderr, we just
stop using them and manipulate environ directly.

Then we arrive at the third problem, FreeBSD is not POSIX-compatible and
you cannot use environ from shared libraries:
https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=263265

So this commit will save a pointer to environ from the executable (not a
shared library) and use this indirection (from the shared library) to
access environ. After secure erasure, environ is set to an empty
environment. Then only after the fork, we set environ to a new
environment dynamically allocated by ourselves (instead of relying on
setenv()).