Change default record_separator in stream.scanner to LF
It was a bad idea to use CRLF as the default. Network protocols are
complex and most commonly abstracted in whole classes.

Meanwhile one hardly ever has specialized parsers to UNIX files that are
used to interact with the OS (e.g. cgroup files). Parsing these files
are usually an ad-hoc operation and having the default that serves these
needs is of more value.