Always start subprocess-based actors with umask 022
There's no point in synchronizing umask values across actors in the
system. Umask is not a credential.

A default value for this process-level property is useful for
predictability which helps in multi-threaded systems (e.g. slave
thread-based actors spawning subprocess-based actors). The value 022 for
umask is a conservative choice that we can safely use as a default.