Make non-IPC actor messaging more asynchronous
Emilua intentionally used lots of synchronization points internally for
actor messaging as it'd be easier to remove synchronization than to add
(if the chosen semantics proved to be wrong later). Implementation-wise,
it's always easier to remove guarantees than to add them.

Fast-forward to the present and it's clear now that the excessive
synchronization is not really useful. The excessive synchronization was
not getting in the way for anything, but it wasn't needed either.

`channel.send` still retains some of the previous properties such as
most of the error-checking (e.g. detecting channel-closed for many
scenarios), post semantics in ASIO-lingo (fiber goes to the end of the
execution queue so other fibers have a chance to run), and fiber
interruptibility.

However not all error checking that was available previously remains as
it's impossible to implement all of them in an asynchronous system (by
the time an error is detected such as channel-closed the sending actor
already moved on to other tasks and we cannot travel back to the past to
make `channel.send` raise an exception there). Some test cases were
modified to cope with the removal of these guarantees.

We could go further and just don't reschedule the fiber nor check for
interruptions at all, but I feel more comfortable doing small gradual
changes to see how the changes play out over time.

The new semantics also make IPC actors and non-IPC actors closer to each
other in behavior.

The new semantics (`channel.send` is fully asynchronous to the target
actor) are lighter to implement as well so it might benefit some
workloads.