Add a new extensible source for builtin modules
The idea is to allow the user to use libemilua overriding certain
functions such that we query these modules before the cache. The
advantages over pre-filling the modules cache are:

* Performance (can use gperf and save RAM + startup time).
* Modules are available to all children VMs as well (specially important
  for sandboxed self-compartmentalised applications).

As of now, only UNIX platforms can make use of this source. The
mechanism can improved later to support Windows as well.