Refactor documentation pages to include ids on tutorials
This change makes it easier to create sed scripts that convert xref
statements into anchors (id can now be predicted from the filename, and
the filename is what we use in the xref statement).

Sed scripts are important to convert the scattered doc pages into a
single PDF (or ePUB) using all.adoc.