Rename is_block_file() to is_block_device()
Also renames is_character_file() to is_character_device().

Emilua was using a convention that comes from C++17's std::filesystem
(which in turn comes from Boost.Filesystem). This convention is dumb. In
UNIX, everything is a file means that saying something is a file
specifies no further traits from an object. Nobody says "is this a
directory file?" because saying a directory is a file conveys no further
information in the sentence.

The new convention used here comes from Python's pathlib.

Fixes #105