Add workaround to build under libc++ (LLVM)
LLVM hasn't yet caught up to P0718R2 (atomic<weak_ptr<T>>), and was so
far failing to build Emilua. FreeBSD uses the LLVM toolchain as the
default toolchain. Emilua builds for FreeBSD had been using GCC so far,
but that is a problem because all libraries in FreeBSD (e.g. libfmt, Qt)
use the libc++ ABI.

The workaround should be slower, but it isn't in the hot path, so
hopefully it's okay for everyday usage.

I can revert this commit once LLVM catches up to P0718R2.