Ensure the standard streams are valid at process startup
Short example on how to reproduce the behavior this patch tries to
protect from:

  head - </proc/cpuinfo >&-

This needs to be done before we create our sandboxes. The sandboxes can
optionally inherit the standard streams. So if the user opts for this
behavior we cannot afford to leak non-related file descriptors that were
created by the runtime at some early point (and whose allocated numbers
are the same of the standard streams).

These tests matter a lot for suid binaries, and it wouldn't be possible
to implement them from the Lua-side. Therefore it's a good thing for the
runtime to have them here.

Fixes #140