Validate fd type for received actor-mboxes
Ideally skipping this validation shouldn't be a problem because errors
would be caught later at use-site. Unfortunately the history of
non-blocking reads/writes on UNIX is not that gentle and MSG_DONTWAIT (a
reactor API) might only work on sockets. Therefore we ensure the fd type
is a socket so we can use MSG_DONTWAIT at least.

The ideal solution for the problem of blocking reads/writes is to just
use proactors. However not every system has one (e.g. FreeBSD only has
proactors for files in the form of POSIX AIO and falls back to a kqueue
reactor for every other IO need).

https://cr.yp.to/unix/nonblock.html