[libmain] Add create_native_modules() and destroy_native_modules()
These hooks can be used so one doesn't have to initialize native modules
as globals before emilua::app_context. Using these hooks the
initialization order for a native module will be uniform no matter
whether the native module is used as plugin or as a builtin object in
the executable itself.

We could even postpone initialization until asio::io_context is created
as well, but we shouldn't do that. If the native module has a dependency
on this order as well then it's just buggy code. So let's just call
create_native_modules() as soon as emilua::app_context is initialized.