Refactor emilua::main::run() to take app_context& as well
When I was refactoring the Qt plugin to work on pluginless mode, I
learned the value of accessing app_context from run(). In the case of
the Qt plugin, there was a need to create a separate thread to actually
run asio::io_context. The main thread would run
QGuiApplication::exec(). However, there was still a need to call
QGuiApplication::exit() after an event that can only be monitored
accessing app_context: extra_threads_count reaches 0.