Fix argv overrun
Although uncommon, &(char*){NULL} is a valid value for argv (i.e. when
argc is zero).

The bug isn't really important today. However there are plans to make
Emilua able to create suid binaries. So this fix is important for the
future where it'll matter.

According to one commit in OpenDoas, it's possible for argc to be -1,
but I didn't quite understand how's that possible. Anyway, it's cheap to
check for negative numbers as well, so let's go ahead and test.