Validate environment values
Linux allows one to exec() using a "corrupt environment" (i.e. an
environment where env vars might not contain an equals sign) so we need
to check for that as well (it's important if one is using Emilua to
create suid binaries).

This was the last preparation step before I start to create suid
binaries myself.