Add parameter subprocess.pd_daemon to spawn_vm()
The default differs from pd_daemon in system.spawn(), but that's
intentional. The use cases are different.

The implementation won't return with error if PD_DAEMON is not supported
by the current OS. That's also intentional. The following will do the
right thing on Linux even if Linux doesn't support PD_DAEMON:

  spawn_vm{
    subprocess = {
      pd_daemon = true,
    }
  }

Therefore there's no reason to reject this argument if unsupported by
the current OS.

Fixes #160