Remove usage of absolute() in custom versions of get_builtin_module()
We can ignore path::root_name() on Windows as long as we know this path
will be invalid on every drive letter. On Windows, "NUL" is not a valid
directory name:
<https://learn.microsoft.com/en-us/windows/win32/fileio/naming-a-file>.

Indiscriminate use of path::absolute() is discouraged because it'll fail
on capsicum mode (and other tightly sandboxed execution environments).
However on Windows, it just doesn't matter.

Adding path::root_name() to path is the only thing that path::absolute()
was doing. We now remove path::root_name() from the other path we
compare against.