Change default concurrency hint from ASIO_CONCURRENCY_HINT_SAFE to 0
ASIO_CONCURRENCY_HINT_SAFE didn't mean what I thought it meant. The new
table from ASIO docs[1] helped to clarify what it actually means.

The only difference from ASIO_CONCURRENCY_HINT_SAFE (old default) and 0
(new default) is the value for the setting scheduler/concurrency_hint.
The new default does what I wrongly assumed the old default was
doing. So this commit is actually a fix to the guarantees that Emilua
was promising to Lua programmers.

Technically it's breaking behavior given the default is
changing. However given the commit is actually a fix, I'll include it
right now (in the same release series).

[1] https://www.boost.org/doc/libs/1_88_0/doc/html/boost_asio/overview/core/configuration.html#boost_asio.overview.core.configuration.configuration_from_concurrency_hint