Force function calls for hook entries
The hook mechanism used in Emilua (when used as a library) is simple: a
function that "does nothing" and the user can override it (e.g. the
symbol is weak in static builds, and the user can use LD_PRELOAD on
dynamic builds).

On ArchLinux packages, the hook for
ipc_actor.cpp:create_native_modules() wasn't working. ipc_actor.cpp
would never call the user replacement for this function.

I've tried a few GCC attributes, but none of them worked for this use
case:

* noplt
* noinline
* noipa
* externally_visible

Then I just used a cast to volatile and it worked. Now I'm using the
same cast-to-volatile pattern for all hooks. There must be a less
aggressive way to solve this problem, but I'm not aware which way is it.

Fixes #178